Adds WIP refactoring into GamesController from BattleshipsController
