+++ api/controllers/battleships.js
@@ -5,7 +5,7 @@ const { TokenExpiredError } = require("jsonwebtoken");
 // TODO Add points for this game if there is a win condition
 
 // ============ HELPER FUNCTIONS FOR CONCEALMENT: ===================
-const concealedGameView = (populatedGame, viewerID) => {
+const concealedBattleshipsView = (populatedGame, viewerID) => {
   // All concealment occurs in the backend before the final game data is returned, rather than in the frontend display methods, so that players cannot cheat by inspecting the data.
   // If the viewer is not the owner of a board or shipyard, they will get a concealed version of each.
   // If the game is over (finished === true): no concealment occurs and the function returns the original populatedGame.
@@ -57,259 +57,34 @@ const concealedGameView = (populatedGame, viewerID) => {
 
 const BattleshipsController = {
   // ================== METHODS SHARED BY ALL GAMES ============================
-  // This Index method only shows the properties of allgames, and no game boards
+  // Method to fetch all Battleships games
   Index: (req, res) => {
-    // using the select method to hide all gameboard details to prevent cheating.
-    Battleships.find()
-      .select({
-        title: 1,
-        endpoint: 1,
-        playerOne: 1,
-        playerTwo: 1,
-        turn: 1,
-        winner: 1,
-        finished: 1,
-      })
-      .populate("playerOne", "_id username points")
-      .populate("playerTwo", "_id username points")
-      .populate("winner", "_id username points")
-      .exec((err, battleshipsGames) => {
-        if (err) {
-          throw err;
-        }
-        const token = TokenGenerator.jsonwebtoken(req.user_id);
-        res.setHeader("Cache-Control", "no-store, no-cache");
-        res.status(200).json({ games: battleshipsGames, token: token });
-      });
+    GamesController.Index(req, res, Battleships, concealedBattleshipsView);
   },
 
-  // This FindByID method conceals the game according to the viewer prior to returning the game data.
+  // Method to fetch a specific Battleships game by ID
   FindByID: (req, res) => {
-    const battleshipsID = req.params.id;
-    const userID = req.user_id; // userID of the viewer
-    const token = TokenGenerator.jsonwebtoken(req.user_id);
-
-    // ========= 1) Find the game ====================
-    Battleships.findById(battleshipsID)
-      .populate("playerOne", "_id username points")
-      .populate("playerTwo", "_id username points")
-      .populate("winner", "_id username points")
-      .exec((err, game) => {
-        if (err) {
-          throw err;
-        }
-
-        // Game not found
-        if (!game) {
-          return res
-            .status(404)
-            .json({ error: "Game not found", token: token });
-        }
-
-        // ======== 2) Conceal the boards according to who is looking (playerOne, playerTwo, outsider) ============
-        game = concealedGameView(game, userID);
-        res.setHeader("Cache-Control", "no-store, no-cache");
-        res.status(200).json({ game: game, token: token });
-      });
+    GamesController.FindByID(req, res, Battleships, concealedBattleshipsView);
   },
 
+  // Method to create a new Battleships game
   Create: async (req, res) => {
-    const userID = req.user_id;
-    const newBattleships = new Battleships({
-      playerOne: userID,
-    });
-
-    try {
-      const result = await newBattleships.save();
-      const populatedBattleships = await Battleships.populate(result, {
-        path: "playerOne",
-        select: "_id username points",
-      });
-      await Battleships.populate(populatedBattleships, {
-        path: "playerTwo",
-        select: "_id username points",
-      });
-      await Battleships.populate(populatedBattleships, {
-        path: "winner",
-        select: "_id username points",
-      });
-
-      const token = TokenGenerator.jsonwebtoken(req.user_id);
-      res.setHeader("Cache-Control", "no-store, no-cache");
-      res.status(201).json({ token: token, game: populatedBattleships });
-    } catch (error) {
-      console.log("Error in TTT.Create", error);
-      res.status(501).json(error);
-    }
+    GamesController.Create(req, res, Battleships);
   },
 
+  // Method to join a Battleships game
   Join: async (req, res) => {
-    const gameID = req.params.id;
-    const userID = req.user_id;
-    const token = TokenGenerator.jsonwebtoken(req.user_id);
-
-    try {
-      const game = await Battleships.findById(gameID);
-
-      // Game not found error
-      if (!game) {
-        return res.status(404).json({ error: "Game not found", token: token });
-      }
-      // Already joined game error
-      if (
-        userID == game.playerOne ||
-        (game.playerTwo && game.playerTwo == userID)
-      ) {
-        console.log("ERROR: ALREADY IN THIS GAME");
-        return res
-          .status(403)
-          .json({ error: "Already in this game", token: token });
-      }
-      // Game is full error
-      if (game.playerTwo) {
-        console.log("ERROR: GAME ALREADY FULL");
-        return res
-          .status(403)
-          .json({ error: "Game already full.", token: token });
-      } else {
-        // Join game
-        const joinedGame = await Battleships.findOneAndUpdate(
-          { _id: gameID },
-          {
-            $set: { playerTwo: userID },
-          },
-          { new: true }
-        )
-          .populate("playerOne", "_id username points")
-          .populate("playerTwo", "_id username points")
-          .populate("winner", "_id username points");
-
-        // ======== Conceal the boards according to who is looking (playerOne, playerTwo, outsider) ============
-        const concealedGame = concealedGameView(joinedGame, userID);
-        res.setHeader("Cache-Control", "no-store, no-cache");
-        res.status(200).json({ token: token, game: concealedGame });
-      }
-    } catch (error) {
-      console.log("Error in TTT.Join", error);
-      res.status(500).json(error);
-    }
+    GamesController.Join(req, res, Battleships, concealedBattleshipsView);
   },
 
+  // Method to forfeit a Battleships game
   Forfeit: async (req, res) => {
-    try {
-      const sessionUser = req.user_id;
-      const gameID = req.params.id;
-      const token = TokenGenerator.jsonwebtoken(req.user_id);
-      const game = await Battleships.findById(gameID);
-
-      // Game not found error
-      if (!game) {
-        return res.status(404).json({ error: "Game not found", token: token });
-      }
-      // Game aready finished error
-      if (game.finished) {
-        console.log("ERROR: GAME ALREADY FINISHED");
-        return res.status(403).json({
-          error: "Game already finished.",
-          token: token,
-        });
-      }
-      if (!game.playerTwo) {
-        console.log("ERROR: AWAITING PLAYER TWO");
-        return res.status(403).json({
-          error: "Awaiting player two. Please delete instead.",
-          token: token,
-        });
-      }
-
-      // Throw error if sessionUser is not in the game:
-      if (sessionUser != game.playerOne && sessionUser != game.playerTwo) {
-        // needs to be != due to Mongoose datatypes
-        console.log("ERROR: NON-PARTICIPANTS CANNOT FORFEIT");
-        return res.status(403).json({
-          error: "Only players can forfeit the game.",
-          token: token,
-        });
-      }
-      // needs to be ==  not === due to Mongoose datatypes
-      const winner =
-        sessionUser == game.playerOne ? game.playerTwo : game.playerOne;
-
-      const forfeitedGame = await Battleships.findOneAndUpdate(
-        { _id: gameID },
-        {
-          $push: { winner: winner },
-          $set: { finished: true },
-        },
-        { new: true }
-      )
-        .populate("playerOne", "_id username points")
-        .populate("playerTwo", "_id username points")
-        .populate("winner", "_id username points");
-
-      // ======== Conceal the boards according to who is looking (playerOne, playerTwo, outsider) ============
-      const concealedGame = concealedGameView(forfeitedGame, sessionUser);
-      res.setHeader("Cache-Control", "no-store, no-cache");
-      res.status(200).json({ token: token, game: concealedGame });
-    } catch (error) {
-      console.error("Error forfeiting: ", error);
-      res.status(500).json(error);
-    }
+    GamesController.Forfeit(req, res, Battleships, concealedBattleshipsView);
   },
 
+  // Method to delete a Battleships game
   Delete: async (req, res) => {
-    try {
-      const sessionUser = req.user_id;
-      const gameID = req.params.id;
-      const token = TokenGenerator.jsonwebtoken(req.user_id);
-      const game = await Battleships.findById(gameID);
-
-      // Game not found error
-      if (!game) {
-        return res.status(404).json({ error: "Game not found", token: token });
-      }
-      // Throw error if sessionUser is not playerOne (host)
-      if (sessionUser != game.playerOne) {
-        // needs to be != due to Mongoose datatypes
-        console.log("ERROR: ONLY HOSTS CAN DELETE GAMES");
-        return res.status(403).json({
-          error: "Only hosts can delete the game.",
-          token: token,
-        });
-      }
-      // Throw error if game is full (has playerTwo):
-      if (game.playerTwo) {
-        console.log("ERROR: CANNOT DELETE NON-OPEN GAMES");
-        return res.status(403).json({
-          error: "Only games awaiting player Two can be deleted.",
-          token: token,
-        });
-      }
-
-      // Delete the game
-      await Battleships.findByIdAndDelete(gameID);
-
-      // Get the updated game list
-      const updatedGames = await Battleships.find()
-        .select({
-          title: 1,
-          endpoint: 1,
-          playerOne: 1,
-          playerTwo: 1,
-          turn: 1,
-          winner: 1,
-          finished: 1,
-        })
-        .populate("playerOne", "_id username points")
-        .populate("playerTwo", "_id username points")
-        .populate("winner", "_id username points");
-
-      res.setHeader("Cache-Control", "no-store, no-cache");
-      res.status(200).json({ token: token, games: updatedGames });
-    } catch (error) {
-      console.error("Error deleting: ", error);
-      res.status(500).json(error);
-    }
+    GamesController.Delete(req, res, Battleships, concealedBattleshipsView);
   },
 
   // ===================== BATTLESHIP SPECIFIC GAMEPLAY METHODS ============================
@@ -436,7 +211,7 @@ const BattleshipsController = {
         .populate("winner", "_id username points");
 
       // ======== 3) Conceal the boards according to who is looking (playerOne, playerTwo, outsider) ============
-      const concealedPlacedShipsGame = concealedGameView(
+      const concealedPlacedShipsGame = concealedBattleshipsView(
         placedShipsGame,
         userID
       );
@@ -627,7 +402,7 @@ const BattleshipsController = {
         .populate("playerTwo", "_id username points")
         .populate("winner", "_id username points");
 
-      const concealedGame = concealedGameView(updatedGame, userID);
+      const concealedGame = concealedBattleshipsView(updatedGame, userID);
       res.setHeader("Cache-Control", "no-store, no-cache");
       res.status(200).json({
         game: concealedGame,
